Auto configuration fine tuning
With some class loaders, the change occurs that if spring-boot/kafka is
on the classpath, that a Spring Boot application is unable to start up
due to a bean name clash on the kafkaConsumerFactory. To that end,
specify Axon's bean names by adding "axon..." to the name. Lastly,
reorder the bean constructors from converter, producers and consumers