Change Fetcher API to be non-streaming centric
The Fetcher API should no longer talk about tracking concepts as it
should be usable for subscribing as well. To that end, remove the start
function and replace it for `poll` with a Consumer, RecordConverter and
ReconrdConsumer given to it. The return value should be a Runnable to
close the fetching process. Consequently the AsyncFetcher implementation
and it's test case should be adjusted

#17