Several fine tuning actions
Firstly, the generic order should be changed from event-key-value, to
key-value-event, as that's the message retrieval, conversion and
consumption order. Secondly, instead of a Runnable, we should return a
Registration to cancel a Fetcher operation. Lastly, the RecordConsumer
should be called EventConsumer to not confuse to much with Kafka's
ConsumerRecord

#17