Introduce a TrackingTokenConsumerRebalanceListener
Introduce a ConsumerRebalanceListener tied to the Consumer's subscribe
call for the streamable solution. This listener will use the assigned
partitions for the given Consumer and check whether the provided
KafkaTrackingToken has an offset for them. If yes, it will call
Consumer$#seek for the provided partition-offset pair. If a new
partition is assigned to the Consumer, thus the KafkaTrackingToken does
not know of it, then we forcefully set the offset to zero to start from
the beginning in that partition. Lastly, to cover bug report #25, we
will use a Token Supplier instead of a Token directly to cover
reassignments in a running system. This will replace the Consumerutil
implementation.

#18 and #25